edited readme and controller functions

+++ routes/api/tag-routes.js
@@ -2,70 +2,68 @@ const { Router } = require("express");
 
 const { Tag, Product, ProductTag } = require("../../models");
 
-// The `/api/tags` endpoint
-
 const router = Router();
-
+// get all tags
 router.get("/", async (req, res) => {
-  // find all tags
-  // be sure to include its associated Product data
   try {
-    const getTags = await Tag.findAll();
-    res.status(200).json(getTags);
+    const getAllTags = await Tag.findAll();
+    res.status(200).json(getAllTags);
   } catch (err) {
     res.status(500).json(err);
   }
 });
-
+// get one tag by its `id`
 router.get("/:id", async (req, res) => {
-  // find a single tag by its `id`
-  // be sure to include its associated Product data
   try {
-    const getTag = await Tag.findByPk(req.params.id, {
-      // JOIN with travellers, using the Trip through table
-      include: [{ model: Product, ProductTag, as: "tag_data" }],
+    const getTagById = await Tag.findByPk(req.params.id, {
+      include: [
+        {
+          model: Product,
+          attributes: ["id", "product_name", "price", "stock", "category_id"],
+        },
+      ],
     });
 
-    if (!getTag) {
-      res.status(404).json({ message: "No tags found with this id!" });
-      return;
+    if (!getTagById) {
+      return res.status(404).json({ message: "No product with this id" });
     }
 
-    res.status(200).json(getTag);
-  } catch (err) {
-    res.status(500).json(err);
+    return res.json(getTagById);
+  } catch (error) {
+    console.error(`[ERROR]: Failed to get tag | ${error.message}`);
+    return res.status(500).json({ error: "Failed to get tag" });
   }
 });
-
+// create a new tag
 router.post("/", async (req, res) => {
-  // create a new tag
   try {
-    const createTag = await Tag.create(req.body);
-    res.status(200).json(createTag);
-  } catch (err) {
-    res.status(400).json(err);
+    const newTag = await Tag.create(req.body);
+    return res.json(newTag);
+  } catch (error) {
+    console.error(`[ERROR]: Failed to create tag | ${error.message}`);
+    return res.status(500).json({ error: "Failed to create tag" });
   }
 });
 
-router.put("/:id", (req, res) => {
-  // update a tag's name by its `id` value
-});
-
+// delete a tag by id
 router.delete("/:id", async (req, res) => {
-  // delete on tag by its `id` value
   try {
-    const deleteTag = await Tag.destroy({
+    const tag = await Tag.findByPk(req.params.id);
+
+    if (!tag) {
+      return res.status(404).json({ message: "No tag found" });
+    }
+
+    await Tag.destroy({
       where: {
         id: req.params.id,
       },
     });
-    if (!deleteTag) {
-      res.status(404).json({ message: "No tags found with this id" });
-      return;
-    }
-    res.status(200).json(deleteTag);
-  } catch (err) {
-    res.status(500).json(err);
+
+    return res.json({ message: "Successfully deleted tag" });
+  } catch (error) {
+    console.error(`[ERROR]: Failed to delete tag | ${error.message}`);
+    return res.status(500).json({ error: "Failed to delete tag" });
   }
 });
